requirements.txt
Django==1.10.1
argparse==1.2.1
django-cors-headers==1.1.0
django-extensions==1.7.4
djangorestframework==3.4.7
psycopg2==2.6.2
six==1.10.0
wsgiref==0.1.2
